Fix scrolling with region
Previously, grid_position_to_doc_position was called once to get the range to delete, then again to get the range to insert (after the deletion was performed). However, the delete reduced the number of lines in the doc, causing grid_position_to_doc_position to return the wrong value the second time it was called. Now we just do our position calculations for both the insertion and deletion at once, then perform the actual deletion and insertion using those ranges.